chore: move all PDF functions to external service
