+++ API/CCW.Application/Controllers/PermitApplicationController.cs
@@ -5,15 +5,9 @@ using CCW.Application.Enum;
 using CCW.Application.Models;
 using CCW.Application.Services.Contracts;
 using CCW.Common.Models;
-using iText.Forms;
-using iText.IO.Image;
-using iText.Kernel.Pdf;
-using iText.Layout;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System.Drawing;
 using System.Globalization;
-using System.Text;
 
 namespace CCW.Application.Controllers;
 
@@ -22,8 +16,6 @@ namespace CCW.Application.Controllers;
 public class PermitApplicationController : ControllerBase
 {
     private readonly IDocumentServiceClient _documentHttpClient;
-    private readonly IAdminServiceClient _adminHttpClient;
-    private readonly IUserProfileServiceClient _userProfileServiceClient;
     private readonly ICosmosDbService _cosmosDbService;
     private readonly IPdfService _pdfService;
     private readonly ILogger<PermitApplicationController> _logger;
@@ -31,8 +23,6 @@ public class PermitApplicationController : ControllerBase
 
     public PermitApplicationController(
         IDocumentServiceClient documentHttpClient,
-        IAdminServiceClient adminHttpClient,
-        IUserProfileServiceClient userProfileServiceClient,
         ICosmosDbService cosmosDbService,
         IPdfService pdfService,
         ILogger<PermitApplicationController> logger,
@@ -40,8 +30,6 @@ public class PermitApplicationController : ControllerBase
         )
     {
         _documentHttpClient = documentHttpClient;
-        _adminHttpClient = adminHttpClient;
-        _userProfileServiceClient = userProfileServiceClient;
         _cosmosDbService = cosmosDbService ?? throw new ArgumentNullException(nameof(cosmosDbService));
         _pdfService = pdfService;
         _logger = logger;
@@ -807,14 +795,6 @@ public class PermitApplicationController : ControllerBase
         }
     }
 
-    private static string BuildApplicantFullName(PermitApplication userApplication)
-    {
-        return (userApplication.Application.PersonalInfo?.FirstName + " " +
-                                   userApplication.Application.PersonalInfo?.MiddleName + " " +
-                                   userApplication.Application.PersonalInfo?.LastName + " " +
-                                   userApplication.Application.PersonalInfo?.Suffix).Trim();
-    }
-
     [Authorize(Policy = "AADUsers")]
     [Route("printUnofficialLicense")]
     [HttpPut]
@@ -828,105 +808,14 @@ public class PermitApplicationController : ControllerBase
             {
                 return NotFound("Permit application cannot be found.");
             }
-            var adminResponse = await _adminHttpClient.GetAgencyProfileSettingsAsync(cancellationToken: default);
-            var response = await _documentHttpClient.GetUnofficialLicenseTemplateAsync(cancellationToken: default);
-
-            Stream streamToReadFrom = await response.Content.ReadAsStreamAsync();
-            MemoryStream outStream = new MemoryStream();
-
-            PdfReader pdfReader = new PdfReader(streamToReadFrom);
-            PdfWriter pdfWriter = new PdfWriter(outStream);
-            PdfDocument doc = new PdfDocument(pdfReader, pdfWriter);
-
-            Document docFileAll = new Document(doc);
-            pdfWriter.SetCloseStream(false);
-
-            PdfAcroForm form = PdfAcroForm.GetAcroForm(doc, true);
-            form.SetGenerateAppearance(true);
-
-            await AddApplicantSignatureImageForUnOfficial(userApplication, docFileAll);
-            await AddApplicantThumbprintImageForUnOfficial(userApplication, docFileAll);
-            await AddApplicantPhotoImageForUnOfficial(userApplication, docFileAll);
-            await AddSheriffLogoForUnOfficial(docFileAll);
-            await AddSheriffIssuingOfficierSignatureImageForUnOfficial(docFileAll);
-
-            form.GetField("AGENCY_NAME").SetValue(adminResponse.AgencyName ?? "", true);
-            form.GetField("AGENCY_ORI").SetValue(adminResponse.ORI ?? "", true);
-            form.GetField("LOCAL_AGENCY_NUMBER").SetValue(adminResponse.LocalAgencyNumber ?? "", true);
-            string fullname = BuildApplicantFullName(userApplication);
-            form.GetField("APPLICANT_NAME").SetValue(fullname.Trim(), true);
-
-            string residenceAddress1 = userApplication.Application.CurrentAddress?.AddressLine1;
-            string residenceAddress2 = userApplication.Application.CurrentAddress?.AddressLine2;
-            if (residenceAddress2 != null)
-            {
-                residenceAddress1 = residenceAddress1 + ", " + residenceAddress2;
-            }
-            form.GetField("APPLICATION_ADDRESS_LINE_1").SetValue(residenceAddress1 ?? "", true);
-            string residenceAddress3 = userApplication.Application.CurrentAddress?.City
-                                       + ", " + userApplication.Application.CurrentAddress?.State
-                                       + " " + userApplication.Application.CurrentAddress?.Zip;
-            form.GetField("APPLICATION_ADDRESS_LINE_2").SetValue(residenceAddress3 ?? "", true);
-            string licenseType = userApplication.Application.ApplicationType?.ToString();
-            licenseType = char.ToUpper(licenseType[0]) + licenseType.Substring(1);
-            form.GetField("LICENSE_TYPE").SetValue(licenseType ?? "", true);
-            form.GetField("DATE_OF_BIRTH").SetValue(userApplication.Application.DOB?.BirthDate ?? "", true);
-            form.GetField("ISSUED_DATE").SetValue(userApplication.Application.License?.IssueDate ?? "", true);
-            form.GetField("EXPIRED_DATE").SetValue(userApplication.Application.License?.ExpirationDate ?? "", true);
-
-            string height = userApplication.Application.PhysicalAppearance?.HeightFeet + "'" + userApplication.Application.PhysicalAppearance?.HeightInch;
-            form.GetField("HEIGHT").SetValue(height ?? "", true);
-            form.GetField("WEIGHT").SetValue(userApplication.Application.PhysicalAppearance?.Weight ?? "", true);
-            form.GetField("EYE_COLOR").SetValue(userApplication.Application.PhysicalAppearance?.EyeColor ?? "", true);
-            form.GetField("HAIR_COLOR").SetValue(userApplication.Application.PhysicalAppearance?.HairColor ?? "", true);
-
-            var weapons = userApplication.Application.Weapons;
-            if (null != weapons && weapons.Length > 0)
-            {
-                int totalWeapons = weapons.Length;
-                int fieldIteration = 1;
-                string makeField;
-                string modelField;
-                string serialField;
-                string caliberField;
-
-                for (int i = 0; i < totalWeapons && i < 4; i++)
-                {
-                    makeField = "MANUFACTURER" + fieldIteration.ToString();
-                    modelField = "MODEL" + fieldIteration.ToString();
-                    serialField = "SERIAL" + fieldIteration.ToString();
-                    caliberField = "CALIBER" + fieldIteration.ToString();
-
-                    form.GetField(makeField).SetValue(weapons[i].Make);
-                    form.GetField(modelField).SetValue(weapons[i].Model);
-                    form.GetField(serialField).SetValue(weapons[i].SerialNumber);
-                    form.GetField(caliberField).SetValue(weapons[i].Caliber);
-
-                    fieldIteration++;
-                }
-            }
-
-            form.GetField("ISSUING_NAME").SetValue(adminResponse.AgencySheriffName ?? "", true);
-            form.GetField("INFO_NUMBER").SetValue(adminResponse.AgencyTelephone ?? "", true);
-            docFileAll.Flush();
-            form.FlattenFields();
-            docFileAll.Close();
-
-            FileStreamResult fileStreamResult = new FileStreamResult(outStream, "application/pdf");
-            FormFile fileToSave = new FormFile(fileStreamResult.FileStream, 0, outStream.Length, null!, fileName);
-
-            var saveFileResult = await _documentHttpClient.SaveAdminApplicationPdfAsync(fileToSave, fileName, cancellationToken: default);
 
             Response.Headers.Append("Content-Disposition", "inline");
             Response.Headers.Add("X-Content-Type-Options", "nosniff");
 
-            byte[] byteInfo = outStream.ToArray();
-            outStream.Write(byteInfo, 0, byteInfo.Length);
-            outStream.Position = 0;
+            var outStream = await _pdfService.GetUnofficialLicenseMemoryStream(userApplication, fileName);
 
             FileStreamResult fileStreamResultDownload = new FileStreamResult(outStream, "application/pdf");
 
-            //Uncomment this to return the file as a download
             if (shouldAddDownloadFilename)
             {
                 fileStreamResultDownload.FileDownloadName = DateTime.Now.ToString("yyyy-MM-dd-HH-mm-ss-") + fileName + ".pdf";
@@ -956,100 +845,14 @@ public class PermitApplicationController : ControllerBase
             {
                 return NotFound("Permit application cannot be found.");
             }
-            var adminResponse = await _adminHttpClient.GetAgencyProfileSettingsAsync(cancellationToken: default);
-            var response = await _documentHttpClient.GetLiveScanTemplateAsync(cancellationToken: default);
-
-            Stream streamToReadFrom = await response.Content.ReadAsStreamAsync();
-            MemoryStream outStream = new MemoryStream();
-
-            PdfReader pdfReader = new PdfReader(streamToReadFrom);
-            PdfWriter pdfWriter = new PdfWriter(outStream);
-            PdfDocument doc = new PdfDocument(pdfReader, pdfWriter);
-
-            Document docFileAll = new Document(doc);
-            pdfWriter.SetCloseStream(false);
-
-            PdfAcroForm form = PdfAcroForm.GetAcroForm(doc, true);
-            form.SetGenerateAppearance(true);
-
-            await AddApplicantSignatureImageForLiveScan(userApplication, docFileAll);
-            var submittedDate = DateTime.Now.ToString("MM/dd/yyyy");
-            form.GetField("DATE").SetValue(submittedDate ?? "", true);
-            form.GetField("ORI").SetValue(adminResponse.ORI ?? "", true);
-            string licenseType = userApplication.Application.ApplicationType?.ToString();
-            licenseType = licenseType.ToUpper() + " CCW";
-            form.GetField("AUTHORIZED_APPLICANT_TYPE").SetValue(licenseType ?? "", true);
-            form.GetField("LICENSE_TYPE").SetValue(licenseType ?? "", true);
-            form.GetField("AGENCY_NAME").SetValue(adminResponse.AgencyName ?? "", true);
-            form.GetField("AGENCY_MAIL_CODE").SetValue(adminResponse.MailCode ?? "", true);
-            form.GetField("AGENCY_ADDRESS_1").SetValue(adminResponse.AgencyShippingStreetAddress ?? "", true);
-            form.GetField("AGENCY_CONTACT_NAME").SetValue(adminResponse.ContactName ?? "", true);
-            form.GetField("AGENCY_CITY").SetValue(adminResponse.AgencyShippingCity ?? "", true);
-            form.GetField("AGENCY_STATE").SetValue(GetStateByName(adminResponse.AgencyShippingState) ?? "", true);
-            form.GetField("AGENCY_ZIP").SetValue(adminResponse.AgencyShippingZip ?? "", true);
-            form.GetField("AGENCY_CONTACT_NUMBER").SetValue(adminResponse.ContactNumber ?? "", true);
-            string fullname = BuildApplicantFullName(userApplication);
-            form.GetField("LAST_NAME").SetValue(userApplication.Application.PersonalInfo?.LastName ?? "", true);
-            form.GetField("FIRST_NAME").SetValue(userApplication.Application.PersonalInfo?.FirstName ?? "", true);
-            if (userApplication.Application.PersonalInfo?.MiddleName != "" && userApplication.Application.PersonalInfo?.MiddleName != null)
-            {
-                form.GetField("MIDDLE_INITIAL").SetValue(userApplication.Application.PersonalInfo?.MiddleName.Substring(0, 1) ?? "", true);
-            }
-
-            form.GetField("SUFFIX").SetValue(userApplication.Application.PersonalInfo?.Suffix ?? "", true);
-            if (userApplication.Application.Aliases.Length > 0)
-            {
-                form.GetField("LAST_NAME_2").SetValue(userApplication.Application.Aliases[0].PrevLastName ?? "", true);
-                form.GetField("FIRST_NAME_2").SetValue(userApplication.Application.Aliases[0].PrevFirstName ?? "", true);
-                form.GetField("SUFFIX_2").SetValue(userApplication.Application.Aliases[0].PrevSuffix ?? "", true);
-            }
-            form.GetField("DATE_OF_BIRTH").SetValue(userApplication.Application.DOB.BirthDate ?? "", true);
-            if (userApplication.Application.PhysicalAppearance.Gender == "male")
-            {
-                form.GetField("MALE").SetValue("true");
-            }
-            else
-            {
-                form.GetField("FEMALE").SetValue("true");
-            }
-            form.GetField("DL_NUMBER").SetValue(userApplication.Application.IdInfo.IdNumber ?? "", true);
-            string height = userApplication.Application.PhysicalAppearance?.HeightFeet + "'" + userApplication.Application.PhysicalAppearance?.HeightInch;
-            form.GetField("HEIGHT").SetValue(height ?? "", true);
-            form.GetField("WEIGHT").SetValue(userApplication.Application.PhysicalAppearance.Weight ?? "", true);
-            form.GetField("EYE_COLOR").SetValue(userApplication.Application.PhysicalAppearance.EyeColor ?? "", true);
-            form.GetField("HAIR_COLOR").SetValue(userApplication.Application.PhysicalAppearance.HairColor ?? "", true);
-            form.GetField("AGENCY_BILLING_NUMBER").SetValue(adminResponse.AgencyBillingNumber ?? "", true);
-            form.GetField("BIRTH_STATE").SetValue(GetStateByName(userApplication.Application.DOB.BirthState) ?? "", true);
-            form.GetField("SSN").SetValue(userApplication.Application.PersonalInfo.Ssn ?? "", true);
-            string residenceAddress1 = userApplication.Application.CurrentAddress?.AddressLine1;
-            string residenceAddress2 = userApplication.Application.CurrentAddress?.AddressLine2;
-            if (residenceAddress2 != null)
-            {
-                residenceAddress1 = residenceAddress1 + ", " + residenceAddress2;
-            }
-            form.GetField("ADDRESS_1").SetValue(residenceAddress1 ?? "", true);
-            form.GetField("CITY").SetValue(userApplication.Application.CurrentAddress?.City ?? "", true);
-            form.GetField("STATE").SetValue(GetStateByName(userApplication.Application.CurrentAddress?.State) ?? "", true);
-            form.GetField("ZIP").SetValue(userApplication.Application.CurrentAddress?.Zip ?? "", true);
-            docFileAll.Flush();
-            form.FlattenFields();
-            docFileAll.Close();
-
-            FileStreamResult fileStreamResult = new FileStreamResult(outStream, "application/pdf");
-            FormFile fileToSave = new FormFile(fileStreamResult.FileStream, 0, outStream.Length, null!, fileName);
-
-            var saveFileResult = await _documentHttpClient.SaveAdminApplicationPdfAsync(fileToSave, fileName, cancellationToken: default);
 
             Response.Headers.Append("Content-Disposition", "inline");
             Response.Headers.Add("X-Content-Type-Options", "nosniff");
 
-            byte[] byteInfo = outStream.ToArray();
-            outStream.Write(byteInfo, 0, byteInfo.Length);
-            outStream.Position = 0;
+            var outStream = await _pdfService.GetLivescanMemoryStream(userApplication, fileName);
 
             FileStreamResult fileStreamResultDownload = new FileStreamResult(outStream, "application/pdf");
 
-            //Uncomment this to return the file as a download
             if (shouldAddDownloadFilename)
             {
                 fileStreamResultDownload.FileDownloadName = DateTime.Now.ToString("yyyy-MM-dd-HH-mm-ss-") + fileName + ".pdf";
@@ -1066,292 +869,9 @@ public class PermitApplicationController : ControllerBase
         }
     }
 
-    private async Task AddApplicantSignatureImageForLiveScan(PermitApplication userApplication, Document docFileAll)
-    {
-        var signatureFileName = BuildApplicantDocumentName(userApplication, "signature");
-        var documentResponse = await _documentHttpClient.GetApplicantImageAsync(signatureFileName, cancellationToken: default);
-        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
-
-        var sr = new StreamReader(streamContent);
-        string imageUri = sr.ReadToEnd();
-        string imageBase64Data = imageUri.Remove(0, 22);
-        byte[] imageBinaryData = Convert.FromBase64String(imageBase64Data);
-
-        var imageData = ImageDataFactory.Create(imageBinaryData);
-        var position = new ImagePosition()
-        {
-            Page = 1,
-            Width = 250,
-            Height = 30,
-            Left = 150,
-            Bottom = 280
-        };
-
-        var image = GetImageForImageData(imageData, position);
-        docFileAll.Add(image);
-    }
-
-    private async Task AddApplicantSignatureImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
-    {
-        var signatureFileName = BuildApplicantDocumentName(userApplication, "signature");
-        var imageData = await GetImageDataForPdf(signatureFileName);
-
-        var leftPosition = new ImagePosition()
-        {
-            Page = 1,
-            Width = 210,
-            Height = 30,
-            Left = 145,
-            Bottom = -6
-        };
-
-        var leftImage = GetImageForImageData(imageData, leftPosition);
-        docFileAll.Add(leftImage);
-    }
-    private async Task AddApplicantThumbprintImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
-    {
-        var signatureFileName = BuildApplicantDocumentName(userApplication, "thumbprint");
-        var imageData = await GetImageDataForPdf(signatureFileName);
-
-        var leftPosition = new ImagePosition()
-        {
-            Page = 2,
-            Width = 40,
-            Height = 50,
-            Left = 181,
-            Bottom = 6
-        };
-
-        var leftImage = GetImageForImageData(imageData, leftPosition);
-        docFileAll.Add(leftImage);
-    }
-    private async Task AddApplicantPhotoImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
-    {
-        var signatureFileName = BuildApplicantDocumentName(userApplication, "portrait");
-        var imageData = await GetImageDataForPdf(signatureFileName);
-
-        var leftPosition = new ImagePosition()
-        {
-            Page = 1,
-            Width = 80,
-            Height = 70,
-            Left = 6,
-            Bottom = 50
-        };
-
-        var leftImage = GetImageForImageData(imageData, leftPosition);
-        docFileAll.Add(leftImage);
-    }
-    private async Task AddSheriffLogoForUnOfficial(Document docFileAll)
-    {
-        var documentResponse = await _documentHttpClient.GetSheriffLogoAsync(cancellationToken: default);
-        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
-
-        var memoryStream = new MemoryStream();
-        await streamContent.CopyToAsync(memoryStream);
-        memoryStream.Position = 0;
-        var imageData = ImageDataFactory.Create(memoryStream.ToArray());
-
-        var leftPosition = new ImagePosition()
-        {
-            Page = 2,
-            Width = 50,
-            Height = 60,
-            Left = 92,
-            Bottom = 6
-        };
-
-        var leftImage = GetImageForImageData(imageData, leftPosition);
-        docFileAll.Add(leftImage);
-    }
-    private async Task AddSheriffIssuingOfficierSignatureImageForUnOfficial(Document docFileAll)
-    {
-        var documentResponse = await _documentHttpClient.GetSheriffSignatureAsync(cancellationToken: default);
-        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
-
-        var memoryStream = new MemoryStream();
-        await streamContent.CopyToAsync(memoryStream);
-        memoryStream.Position = 0;
-        var imageData = ImageDataFactory.Create(memoryStream.ToArray());
-
-        var leftPosition = new ImagePosition()
-        {
-            Page = 2,
-            Width = 180,
-            Height = 17,
-            Left = 2,
-            Bottom = 15
-        };
-
-        var leftImage = GetImageForImageData(imageData, leftPosition);
-        docFileAll.Add(leftImage);
-    }
-
-    private async Task<ImageData> GetImageDataForPdf(string fileName, Stream contentStream = null, bool shouldResize = false)
-    {
-        byte[] imageBinaryData;
-        if (contentStream != null)
-        {
-            var ms = new MemoryStream();
-            await contentStream.CopyToAsync(ms);
-            imageBinaryData = ms.ToArray();
-        }
-        else
-        {
-            var documentResponse = await _documentHttpClient.GetApplicantImageAsync(fileName, cancellationToken: default);
-            imageBinaryData = await documentResponse.Content.ReadAsByteArrayAsync();
-        }
-
-        if (shouldResize)
-        {
-            try
-            {
-                // Ignore these warnings. Technically System.Drawing.Common is NOT cross platform
-                // However, runtimeconfig.template.json setting "System.Drawing.EnableUnixSupport": true
-                // Allows it work on Linux (kind of)
-                System.Drawing.Image image = System.Drawing.Image.FromStream(new MemoryStream(imageBinaryData));
-                Bitmap bmp = new Bitmap(new MemoryStream(imageBinaryData));
-                var resized = ResizeImage(bmp);
-                MemoryStream resizedImageStream = new MemoryStream();
-                resized.Save(resizedImageStream, System.Drawing.Imaging.ImageFormat.Bmp);
-
-                imageBinaryData = resizedImageStream.GetBuffer();
-            }
-            catch (Exception exception)
-            {
-                _logger.LogError(exception, "Error converting image");
-                Console.WriteLine($"Error converting image: {exception.Message}");
-            }
-        }
-
-        var imageData = ImageDataFactory.Create(imageBinaryData);
-
-        return imageData;
-
-        throw new FileNotFoundException("File not found: " + fileName);
-    }
-
-    private Bitmap ResizeImage(Bitmap bitmapToResize)
-    {
-        int w = bitmapToResize.Width;
-        int h = bitmapToResize.Height;
-
-        Func<int, bool> allWhiteRow = row =>
-        {
-            for (int i = 0; i < w; ++i)
-                if (bitmapToResize.GetPixel(i, row).R != 255)
-                    return false;
-            return true;
-        };
-
-        Func<int, bool> allWhiteColumn = col =>
-        {
-            for (int i = 0; i < h; ++i)
-                if (bitmapToResize.GetPixel(col, i).R != 255)
-                    return false;
-            return true;
-        };
-
-        int topmost = 0;
-
-        for (int row = 0; row < h; ++row)
-        {
-            if (allWhiteRow(row))
-                topmost = row;
-            else break;
-        }
-
-        int bottommost = 0;
-        for (int row = h - 1; row >= 0; --row)
-        {
-            if (allWhiteRow(row))
-                bottommost = row;
-            else break;
-        }
-
-        int leftmost = 0, rightmost = 0;
-        for (int col = 0; col < w; ++col)
-        {
-            if (allWhiteColumn(col))
-                leftmost = col;
-            else
-                break;
-        }
-
-        for (int col = w - 1; col >= 0; --col)
-        {
-            if (allWhiteColumn(col))
-                rightmost = col;
-            else
-                break;
-        }
-
-        if (rightmost == 0) rightmost = w; // As reached left
-        if (bottommost == 0) bottommost = h; // As reached top.
-
-        int croppedWidth = rightmost - leftmost;
-        int croppedHeight = bottommost - topmost;
-
-        if (croppedWidth == 0) // No border on left or right
-        {
-            leftmost = 0;
-            croppedWidth = w;
-        }
-
-        if (croppedHeight == 0) // No border on top or bottom
-        {
-            topmost = 0;
-            croppedHeight = h;
-        }
-
-        try
-        {
-            var target = new Bitmap(croppedWidth, croppedHeight);
-            using (Graphics g = Graphics.FromImage(target))
-            {
-                g.DrawImage(bitmapToResize,
-                  new System.Drawing.RectangleF(0, 0, croppedWidth, croppedHeight),
-                  new System.Drawing.RectangleF(leftmost, topmost, croppedWidth, croppedHeight),
-                  GraphicsUnit.Pixel);
-            }
-            return target;
-        }
-        catch (Exception ex)
-        {
-            throw new Exception(
-              string.Format("Values are topmost={0} btm={1} left={2} right={3} croppedWidth={4} croppedHeight={5}", topmost, bottommost, leftmost, rightmost, croppedWidth, croppedHeight),
-              ex);
-        }
-    }
-
-    private iText.Layout.Element.Image GetImageForImageData(ImageData imageData, ImagePosition imagePosition)
-    {
-        return new iText.Layout.Element.Image(imageData)
-            .ScaleToFit(imagePosition.Width, imagePosition.Height)
-            .SetFixedPosition(imagePosition.Page, imagePosition.Left, imagePosition.Bottom);
-    }
-
-    private string BuildApplicantDocumentName(PermitApplication? userApplication, string documentName)
-    {
-        string fullFilename = userApplication.UserId + "_" +
-            userApplication.Application.PersonalInfo?.LastName + "_" +
-            userApplication.Application.PersonalInfo?.FirstName + "_" + documentName;
-
-        return fullFilename;
-    }
-
-    private sealed class ImagePosition
-    {
-        public int Page { get; set; }
-        public int Width { get; set; }
-        public int Height { get; set; }
-        public int Left { get; set; }
-        public int Bottom { get; set; }
-    }
-
     private void GetAADUserName(out string userName)
     {
-        userName = this.HttpContext.User.Claims
+        userName = HttpContext.User.Claims
             .Where(c => c.Type == "preferred_username").Select(c => c.Value)
             .FirstOrDefault();
 
@@ -1370,197 +890,6 @@ public class PermitApplicationController : ControllerBase
         if (userId == null)
         {
             throw new ArgumentNullException("userId", "Invalid token.");
-        }
-    }
-
-    private static string GetStateByName(string name)
-    {
-        if (string.IsNullOrEmpty(name))
-        {
-            return string.Empty;
-        }
-
-        switch (name.ToUpper())
-        {
-            case "ALABAMA":
-                return "AL";
-
-            case "ALASKA":
-                return "AK";
-
-            case "AMERICAN SAMOA":
-                return "AS";
-
-            case "ARIZONA":
-                return "AZ";
-
-            case "ARKANSAS":
-                return "AR";
-
-            case "CALIFORNIA":
-                return "CA";
-
-            case "COLORADO":
-                return "CO";
-
-            case "CONNECTICUT":
-                return "CT";
-
-            case "DELAWARE":
-                return "DE";
-
-            case "DISTRICT OF COLUMBIA":
-                return "DC";
-
-            case "FEDERATED STATES OF MICRONESIA":
-                return "FM";
-
-            case "FLORIDA":
-                return "FL";
-
-            case "GEORGIA":
-                return "GA";
-
-            case "GUAM":
-                return "GU";
-
-            case "HAWAII":
-                return "HI";
-
-            case "IDAHO":
-                return "ID";
-
-            case "ILLINOIS":
-                return "IL";
-
-            case "INDIANA":
-                return "IN";
-
-            case "IOWA":
-                return "IA";
-
-            case "KANSAS":
-                return "KS";
-
-            case "KENTUCKY":
-                return "KY";
-
-            case "LOUISIANA":
-                return "LA";
-
-            case "MAINE":
-                return "ME";
-
-            case "MARSHALL ISLANDS":
-                return "MH";
-
-            case "MARYLAND":
-                return "MD";
-
-            case "MASSACHUSETTS":
-                return "MA";
-
-            case "MICHIGAN":
-                return "MI";
-
-            case "MINNESOTA":
-                return "MN";
-
-            case "MISSISSIPPI":
-                return "MS";
-
-            case "MISSOURI":
-                return "MO";
-
-            case "MONTANA":
-                return "MT";
-
-            case "NEBRASKA":
-                return "NE";
-
-            case "NEVADA":
-                return "NV";
-
-            case "NEW HAMPSHIRE":
-                return "NH";
-
-            case "NEW JERSEY":
-                return "NJ";
-
-            case "NEW MEXICO":
-                return "NM";
-
-            case "NEW YORK":
-                return "NY";
-
-            case "NORTH CAROLINA":
-                return "NC";
-
-            case "NORTH DAKOTA":
-                return "ND";
-
-            case "NORTHERN MARIANA ISLANDS":
-                return "MP";
-
-            case "OHIO":
-                return "OH";
-
-            case "OKLAHOMA":
-                return "OK";
-
-            case "OREGON":
-                return "OR";
-
-            case "PALAU":
-                return "PW";
-
-            case "PENNSYLVANIA":
-                return "PA";
-
-            case "PUERTO RICO":
-                return "PR";
-
-            case "RHODE ISLAND":
-                return "RI";
-
-            case "SOUTH CAROLINA":
-                return "SC";
-
-            case "SOUTH DAKOTA":
-                return "SD";
-
-            case "TENNESSEE":
-                return "TN";
-
-            case "TEXAS":
-                return "TX";
-
-            case "UTAH":
-                return "UT";
-
-            case "VERMONT":
-                return "VT";
-
-            case "VIRGIN ISLANDS":
-                return "VI";
-
-            case "VIRGINIA":
-                return "VA";
-
-            case "WASHINGTON":
-                return "WA";
-
-            case "WEST VIRGINIA":
-                return "WV";
-
-            case "WISCONSIN":
-                return "WI";
-
-            case "WYOMING":
-                return "WY";
-
-            default:
-                return name;
         }
     }
 }
