+++ API/CCW.Application/Services/Contracts/IPdfService.cs
@@ -7,4 +7,6 @@ public interface IPdfService
     public Task<MemoryStream> GetApplicationMemoryStream(PermitApplication userApplication, string licensingUserName, string fileName);
     public Task<MemoryStream> GetRevocationLetterMemoryStream(PermitApplication userApplication, string user, string licensingUserName, string reason, string date, string fileName);
     public Task<MemoryStream> GetOfficialLicenseMemoryStream(PermitApplication userApplication, string licensingUser, string fileName);
+    public Task<MemoryStream> GetUnofficialLicenseMemoryStream(PermitApplication userApplication, string fileName);
+    public Task<MemoryStream> GetLivescanMemoryStream(PermitApplication userApplication, string fileName);
 }
