+++ API/CCW.Application/Services/PdfService.cs
@@ -6,21 +6,18 @@ using iText.IO.Image;
 using iText.Kernel.Pdf;
 using iText.Layout;
 using Microsoft.AspNetCore.Mvc;
-using System.Globalization;
-using System.Text;
-namespace CCW.Application.Services;
-
 using CCW.Application.Enum;
 using iText.Kernel.Colors;
 using iText.Kernel.Geom;
 using iText.Layout.Borders;
-
 using iText.Layout.Element;
 using System.Drawing;
 using System.Globalization;
 using System.Text;
 using System.Text.RegularExpressions;
 using Rectangle = iText.Kernel.Geom.Rectangle;
+
+namespace CCW.Application.Services;
 
 public class PdfService : IPdfService
 {
@@ -30,7 +27,7 @@ public class PdfService : IPdfService
     private readonly ICosmosDbService _cosmosDbService;
 
     public PdfService(
-        IDocumentServiceClient documentHttpClient, 
+        IDocumentServiceClient documentHttpClient,
         IUserProfileServiceClient userProfileServiceClient,
         ICosmosDbService cosmosDbService,
         IAdminServiceClient adminHttpClient)
@@ -887,6 +884,198 @@ public class PdfService : IPdfService
         return outStream;
     }
 
+    public async Task<MemoryStream> GetUnofficialLicenseMemoryStream(PermitApplication userApplication, string fileName)
+    {
+        var adminResponse = await _adminHttpClient.GetAgencyProfileSettingsAsync(cancellationToken: default);
+        var response = await _documentHttpClient.GetUnofficialLicenseTemplateAsync(cancellationToken: default);
+
+        Stream streamToReadFrom = await response.Content.ReadAsStreamAsync();
+        MemoryStream outStream = new MemoryStream();
+
+        PdfReader pdfReader = new PdfReader(streamToReadFrom);
+        PdfWriter pdfWriter = new PdfWriter(outStream);
+        PdfDocument doc = new PdfDocument(pdfReader, pdfWriter);
+
+        Document docFileAll = new Document(doc);
+        pdfWriter.SetCloseStream(false);
+
+        PdfAcroForm form = PdfAcroForm.GetAcroForm(doc, true);
+        form.SetGenerateAppearance(true);
+
+        await AddApplicantSignatureImageForUnOfficial(userApplication, docFileAll);
+        await AddApplicantThumbprintImageForUnOfficial(userApplication, docFileAll);
+        await AddApplicantPhotoImageForUnOfficial(userApplication, docFileAll);
+        await AddSheriffLogoForUnOfficial(docFileAll);
+        await AddSheriffIssuingOfficierSignatureImageForUnOfficial(docFileAll);
+
+        form.GetField("AGENCY_NAME").SetValue(adminResponse.AgencyName ?? "", true);
+        form.GetField("AGENCY_ORI").SetValue(adminResponse.ORI ?? "", true);
+        form.GetField("LOCAL_AGENCY_NUMBER").SetValue(adminResponse.LocalAgencyNumber ?? "", true);
+        string fullname = BuildApplicantFullName(userApplication);
+        form.GetField("APPLICANT_NAME").SetValue(fullname.Trim(), true);
+
+        string residenceAddress1 = userApplication.Application.CurrentAddress?.AddressLine1;
+        string residenceAddress2 = userApplication.Application.CurrentAddress?.AddressLine2;
+        if (residenceAddress2 != null)
+        {
+            residenceAddress1 = residenceAddress1 + ", " + residenceAddress2;
+        }
+        form.GetField("APPLICATION_ADDRESS_LINE_1").SetValue(residenceAddress1 ?? "", true);
+        string residenceAddress3 = userApplication.Application.CurrentAddress?.City
+                                   + ", " + userApplication.Application.CurrentAddress?.State
+                                   + " " + userApplication.Application.CurrentAddress?.Zip;
+        form.GetField("APPLICATION_ADDRESS_LINE_2").SetValue(residenceAddress3 ?? "", true);
+        string licenseType = userApplication.Application.ApplicationType?.ToString();
+        licenseType = char.ToUpper(licenseType[0]) + licenseType.Substring(1);
+        form.GetField("LICENSE_TYPE").SetValue(licenseType ?? "", true);
+        form.GetField("DATE_OF_BIRTH").SetValue(userApplication.Application.DOB?.BirthDate ?? "", true);
+        form.GetField("ISSUED_DATE").SetValue(userApplication.Application.License?.IssueDate ?? "", true);
+        form.GetField("EXPIRED_DATE").SetValue(userApplication.Application.License?.ExpirationDate ?? "", true);
+
+        string height = userApplication.Application.PhysicalAppearance?.HeightFeet + "'" + userApplication.Application.PhysicalAppearance?.HeightInch;
+        form.GetField("HEIGHT").SetValue(height ?? "", true);
+        form.GetField("WEIGHT").SetValue(userApplication.Application.PhysicalAppearance?.Weight ?? "", true);
+        form.GetField("EYE_COLOR").SetValue(userApplication.Application.PhysicalAppearance?.EyeColor ?? "", true);
+        form.GetField("HAIR_COLOR").SetValue(userApplication.Application.PhysicalAppearance?.HairColor ?? "", true);
+
+        var weapons = userApplication.Application.Weapons;
+        if (null != weapons && weapons.Length > 0)
+        {
+            int totalWeapons = weapons.Length;
+            int fieldIteration = 1;
+            string makeField;
+            string modelField;
+            string serialField;
+            string caliberField;
+
+            for (int i = 0; i < totalWeapons && i < 4; i++)
+            {
+                makeField = "MANUFACTURER" + fieldIteration.ToString();
+                modelField = "MODEL" + fieldIteration.ToString();
+                serialField = "SERIAL" + fieldIteration.ToString();
+                caliberField = "CALIBER" + fieldIteration.ToString();
+
+                form.GetField(makeField).SetValue(weapons[i].Make);
+                form.GetField(modelField).SetValue(weapons[i].Model);
+                form.GetField(serialField).SetValue(weapons[i].SerialNumber);
+                form.GetField(caliberField).SetValue(weapons[i].Caliber);
+
+                fieldIteration++;
+            }
+        }
+
+        form.GetField("ISSUING_NAME").SetValue(adminResponse.AgencySheriffName ?? "", true);
+        form.GetField("INFO_NUMBER").SetValue(adminResponse.AgencyTelephone ?? "", true);
+        docFileAll.Flush();
+        form.FlattenFields();
+        docFileAll.Close();
+
+        FileStreamResult fileStreamResult = new FileStreamResult(outStream, "application/pdf");
+        FormFile fileToSave = new FormFile(fileStreamResult.FileStream, 0, outStream.Length, null!, fileName);
+
+        var saveFileResult = await _documentHttpClient.SaveAdminApplicationPdfAsync(fileToSave, fileName, cancellationToken: default);
+
+        byte[] byteInfo = outStream.ToArray();
+        outStream.Write(byteInfo, 0, byteInfo.Length);
+        outStream.Position = 0;
+
+        return outStream;
+    }
+
+    public async Task<MemoryStream> GetLivescanMemoryStream(PermitApplication userApplication, string fileName)
+    {
+        var adminResponse = await _adminHttpClient.GetAgencyProfileSettingsAsync(cancellationToken: default);
+        var response = await _documentHttpClient.GetLiveScanTemplateAsync(cancellationToken: default);
+
+        Stream streamToReadFrom = await response.Content.ReadAsStreamAsync();
+        MemoryStream outStream = new MemoryStream();
+
+        PdfReader pdfReader = new PdfReader(streamToReadFrom);
+        PdfWriter pdfWriter = new PdfWriter(outStream);
+        PdfDocument doc = new PdfDocument(pdfReader, pdfWriter);
+
+        Document docFileAll = new Document(doc);
+        pdfWriter.SetCloseStream(false);
+
+        PdfAcroForm form = PdfAcroForm.GetAcroForm(doc, true);
+        form.SetGenerateAppearance(true);
+
+        await AddApplicantSignatureImageForLiveScan(userApplication, docFileAll);
+        var submittedDate = DateTime.Now.ToString("MM/dd/yyyy");
+        form.GetField("DATE").SetValue(submittedDate ?? "", true);
+        form.GetField("ORI").SetValue(adminResponse.ORI ?? "", true);
+        string licenseType = userApplication.Application.ApplicationType?.ToString();
+        licenseType = licenseType.ToUpper() + " CCW";
+        form.GetField("AUTHORIZED_APPLICANT_TYPE").SetValue(licenseType ?? "", true);
+        form.GetField("LICENSE_TYPE").SetValue(licenseType ?? "", true);
+        form.GetField("AGENCY_NAME").SetValue(adminResponse.AgencyName ?? "", true);
+        form.GetField("AGENCY_MAIL_CODE").SetValue(adminResponse.MailCode ?? "", true);
+        form.GetField("AGENCY_ADDRESS_1").SetValue(adminResponse.AgencyShippingStreetAddress ?? "", true);
+        form.GetField("AGENCY_CONTACT_NAME").SetValue(adminResponse.ContactName ?? "", true);
+        form.GetField("AGENCY_CITY").SetValue(adminResponse.AgencyShippingCity ?? "", true);
+        form.GetField("AGENCY_STATE").SetValue(GetStateByName(adminResponse.AgencyShippingState) ?? "", true);
+        form.GetField("AGENCY_ZIP").SetValue(adminResponse.AgencyShippingZip ?? "", true);
+        form.GetField("AGENCY_CONTACT_NUMBER").SetValue(adminResponse.ContactNumber ?? "", true);
+        string fullname = BuildApplicantFullName(userApplication);
+        form.GetField("LAST_NAME").SetValue(userApplication.Application.PersonalInfo?.LastName ?? "", true);
+        form.GetField("FIRST_NAME").SetValue(userApplication.Application.PersonalInfo?.FirstName ?? "", true);
+        if (userApplication.Application.PersonalInfo?.MiddleName != "" && userApplication.Application.PersonalInfo?.MiddleName != null)
+        {
+            form.GetField("MIDDLE_INITIAL").SetValue(userApplication.Application.PersonalInfo?.MiddleName.Substring(0, 1) ?? "", true);
+        }
+
+        form.GetField("SUFFIX").SetValue(userApplication.Application.PersonalInfo?.Suffix ?? "", true);
+        if (userApplication.Application.Aliases.Length > 0)
+        {
+            form.GetField("LAST_NAME_2").SetValue(userApplication.Application.Aliases[0].PrevLastName ?? "", true);
+            form.GetField("FIRST_NAME_2").SetValue(userApplication.Application.Aliases[0].PrevFirstName ?? "", true);
+            form.GetField("SUFFIX_2").SetValue(userApplication.Application.Aliases[0].PrevSuffix ?? "", true);
+        }
+        form.GetField("DATE_OF_BIRTH").SetValue(userApplication.Application.DOB.BirthDate ?? "", true);
+        if (userApplication.Application.PhysicalAppearance.Gender == "male")
+        {
+            form.GetField("MALE").SetValue("true");
+        }
+        else
+        {
+            form.GetField("FEMALE").SetValue("true");
+        }
+        form.GetField("DL_NUMBER").SetValue(userApplication.Application.IdInfo.IdNumber ?? "", true);
+        string height = userApplication.Application.PhysicalAppearance?.HeightFeet + "'" + userApplication.Application.PhysicalAppearance?.HeightInch;
+        form.GetField("HEIGHT").SetValue(height ?? "", true);
+        form.GetField("WEIGHT").SetValue(userApplication.Application.PhysicalAppearance.Weight ?? "", true);
+        form.GetField("EYE_COLOR").SetValue(userApplication.Application.PhysicalAppearance.EyeColor ?? "", true);
+        form.GetField("HAIR_COLOR").SetValue(userApplication.Application.PhysicalAppearance.HairColor ?? "", true);
+        form.GetField("AGENCY_BILLING_NUMBER").SetValue(adminResponse.AgencyBillingNumber ?? "", true);
+        form.GetField("BIRTH_STATE").SetValue(GetStateByName(userApplication.Application.DOB.BirthState) ?? "", true);
+        form.GetField("SSN").SetValue(userApplication.Application.PersonalInfo.Ssn ?? "", true);
+        string residenceAddress1 = userApplication.Application.CurrentAddress?.AddressLine1;
+        string residenceAddress2 = userApplication.Application.CurrentAddress?.AddressLine2;
+        if (residenceAddress2 != null)
+        {
+            residenceAddress1 = residenceAddress1 + ", " + residenceAddress2;
+        }
+        form.GetField("ADDRESS_1").SetValue(residenceAddress1 ?? "", true);
+        form.GetField("CITY").SetValue(userApplication.Application.CurrentAddress?.City ?? "", true);
+        form.GetField("STATE").SetValue(GetStateByName(userApplication.Application.CurrentAddress?.State) ?? "", true);
+        form.GetField("ZIP").SetValue(userApplication.Application.CurrentAddress?.Zip ?? "", true);
+        docFileAll.Flush();
+        form.FlattenFields();
+        docFileAll.Close();
+
+        FileStreamResult fileStreamResult = new FileStreamResult(outStream, "application/pdf");
+        FormFile fileToSave = new FormFile(fileStreamResult.FileStream, 0, outStream.Length, null!, fileName);
+
+        var saveFileResult = await _documentHttpClient.SaveAdminApplicationPdfAsync(fileToSave, fileName, cancellationToken: default);
+
+
+        byte[] byteInfo = outStream.ToArray();
+        outStream.Write(byteInfo, 0, byteInfo.Length);
+        outStream.Position = 0;
+
+        return outStream;
+    }
+
     private async Task AddApplicantSignatureImageForApplication(PermitApplication userApplication, Document mainDocument)
     {
         string fullFilename = userApplication.UserId + "_" +
@@ -1333,6 +1522,31 @@ public class PdfService : IPdfService
         mainDocument.Add(rightImage);
     }
 
+    private async Task AddApplicantSignatureImageForLiveScan(PermitApplication userApplication, Document docFileAll)
+    {
+        var signatureFileName = BuildApplicantDocumentName(userApplication, "signature");
+        var documentResponse = await _documentHttpClient.GetApplicantImageAsync(signatureFileName, cancellationToken: default);
+        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
+
+        var sr = new StreamReader(streamContent);
+        string imageUri = sr.ReadToEnd();
+        string imageBase64Data = imageUri.Remove(0, 22);
+        byte[] imageBinaryData = Convert.FromBase64String(imageBase64Data);
+
+        var imageData = ImageDataFactory.Create(imageBinaryData);
+        var position = new ImagePosition()
+        {
+            Page = 1,
+            Width = 250,
+            Height = 30,
+            Left = 150,
+            Bottom = 280
+        };
+
+        var image = GetImageForImageData(imageData, position);
+        docFileAll.Add(image);
+    }
+
     private async Task AddApplicantThumbprintImageForOfficial(PermitApplication userApplication, Document mainDocument)
     {
         var signatureFileName = BuildApplicantDocumentName(userApplication, "thumbprint");
@@ -1391,6 +1605,102 @@ public class PdfService : IPdfService
 
         var rightImage = GetImageForImageData(imageData, rightPosition);
         mainDocument.Add(rightImage);
+    }
+
+    private async Task AddApplicantSignatureImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
+    {
+        var signatureFileName = BuildApplicantDocumentName(userApplication, "signature");
+        var imageData = await GetImageDataForPdf(signatureFileName);
+
+        var leftPosition = new ImagePosition()
+        {
+            Page = 1,
+            Width = 210,
+            Height = 30,
+            Left = 145,
+            Bottom = -6
+        };
+
+        var leftImage = GetImageForImageData(imageData, leftPosition);
+        docFileAll.Add(leftImage);
+    }
+    private async Task AddApplicantThumbprintImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
+    {
+        var signatureFileName = BuildApplicantDocumentName(userApplication, "thumbprint");
+        var imageData = await GetImageDataForPdf(signatureFileName);
+
+        var leftPosition = new ImagePosition()
+        {
+            Page = 2,
+            Width = 40,
+            Height = 50,
+            Left = 181,
+            Bottom = 6
+        };
+
+        var leftImage = GetImageForImageData(imageData, leftPosition);
+        docFileAll.Add(leftImage);
+    }
+    private async Task AddApplicantPhotoImageForUnOfficial(PermitApplication userApplication, Document docFileAll)
+    {
+        var signatureFileName = BuildApplicantDocumentName(userApplication, "portrait");
+        var imageData = await GetImageDataForPdf(signatureFileName);
+
+        var leftPosition = new ImagePosition()
+        {
+            Page = 1,
+            Width = 80,
+            Height = 70,
+            Left = 6,
+            Bottom = 50
+        };
+
+        var leftImage = GetImageForImageData(imageData, leftPosition);
+        docFileAll.Add(leftImage);
+    }
+    private async Task AddSheriffLogoForUnOfficial(Document docFileAll)
+    {
+        var documentResponse = await _documentHttpClient.GetSheriffLogoAsync(cancellationToken: default);
+        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
+
+        var memoryStream = new MemoryStream();
+        await streamContent.CopyToAsync(memoryStream);
+        memoryStream.Position = 0;
+        var imageData = ImageDataFactory.Create(memoryStream.ToArray());
+
+        var leftPosition = new ImagePosition()
+        {
+            Page = 2,
+            Width = 50,
+            Height = 60,
+            Left = 92,
+            Bottom = 6
+        };
+
+        var leftImage = GetImageForImageData(imageData, leftPosition);
+        docFileAll.Add(leftImage);
+    }
+    private async Task AddSheriffIssuingOfficierSignatureImageForUnOfficial(Document docFileAll)
+    {
+        var documentResponse = await _documentHttpClient.GetSheriffSignatureAsync(cancellationToken: default);
+        var streamContent = await documentResponse.Content.ReadAsStreamAsync();
+
+        var memoryStream = new MemoryStream();
+        await streamContent.CopyToAsync(memoryStream);
+        memoryStream.Position = 0;
+        var imageData = ImageDataFactory.Create(memoryStream.ToArray());
+
+        var leftPosition = new ImagePosition()
+        {
+            Page = 2,
+            Width = 180,
+            Height = 17,
+            Left = 2,
+            Bottom = 15
+        };
+
+        var leftImage = GetImageForImageData(imageData, leftPosition);
+        docFileAll.Add(leftImage);
     }
 
     private string BuildApplicantDocumentName(PermitApplication? userApplication, string documentName)
